Create a source_generator and build it
The source_generator provides a means to execute the source
sampling outside of an OpenMC simulation. This may be of interest
for diagnostic and testing purposes.

Add the relevant information to the Makefile to build the
source_generator and clean resulting binaries.

Add items to .gitignore so that executable and output are not
picked up.