Add the temperature beta value
The Fausser paper describes the exponent on the ratio of the sampled
radius to the pedestal radius as being a variable beta_T, but the
current implementation fixes this at a value of 2. This commit makes
the beta_T value variable and sets the default value of 6.0 as used in
the paper.