Read C++ and Makefile files into Python file
This commit provides a utility to read the C++ and Makefile files into
a Python file so that the code can be updated via a script when a new
C++ version is available. This is a halfway house to automatically
generating the C++, as it is not possible to create a shared object
library directly from Python.

Also updates .gitignore to ignore compiled Python files.