Install bindings with setup.py
Include instructions in setup.py to compile the bindings with cmake.
Specifically only for the plasma_source library, since we don't need
the source_sampling library in Python.

Update CMakeLists to avoid forcing the library output directory as this
conflicts with the way setuptools etc. look for package artefacts. This
means that the resulting source_sampling library is available in the
build directory, unless the CMAKE_LIBRARY_OUTPUT_DIRECTORY is
passed into the cmake configuration step.

Introduce an __init__.py so that the modules can easily be loaded and
found by setuptools.

Cleanup .gitignore and add dist and egg paths.